code: add utils class
