+++ technologydata/__init__.py
@@ -7,11 +7,14 @@ from technologydata.technologies import (
     Technologies,
     check_source_validity,
 )
+from technologydata.utils import DateFormatEnum, Utils
 
 __version__ = version("technologydata")
 
 
 __all__ = [
+    "Utils",
+    "DateFormatEnum",
     "Source",
     "Sources",
     "Technologies",
