+++ technologydata/sources.py
@@ -5,76 +5,19 @@ import logging
 import pathlib
 import subprocess
 from collections.abc import Iterable
-from enum import Enum
 from typing import Any
 
 import frictionless as ftl
 import pandas as pd
 import requests
 import savepagenow
-from dateutil import parser
+
+import technologydata as td
 
 logger = logging.getLogger(__name__)
 
 DATASOURCES_PATH = pathlib.Path(__file__).parent / "datasources"
 SPECIFICATIONS_PATH = DATASOURCES_PATH / "specification"
-
-
-class DateFormatEnum(str, Enum):
-    """
-    Enum for date formats used in different sources.
-
-    Attributes
-    ----------
-    SOURCES_CSV : str
-        Date format for CSV sources, e.g., "2023-10-01 12:00:00".
-    WAYBACK : str
-        Date format for Wayback Machine, e.g., "20231001120000".
-    NONE : str
-        Represents an empty date format.
-
-    """
-
-    SOURCES_CSV = "%Y-%m-%d %H:%M:%S"
-    WAYBACK = "%Y%m%d%H%M%S"
-    NONE = ""
-
-
-def get_date_format(source: str) -> DateFormatEnum:
-    """
-    Get the date format corresponding to the given source.
-
-    Parameters
-    ----------
-    source : str
-        The source name for which the date format is requested.
-        Accepted values are "sources_csv" and "wayback".
-
-    Returns
-    -------
-    DateFormatEnum
-        The date format associated with the specified source.
-        Returns DateFormatEnum.NONE if the source is not recognized.
-
-    Examples
-    --------
-    >>> get_date_format("sources_csv")
-    <DateFormatEnum.SOURCES_CSV: '%Y-%m-%d %H:%M:%S'>
-
-    >>> get_date_format("wayback")
-    <DateFormatEnum.WAYBACK: '%Y%m%d%H%M%S'>
-
-    >>> get_date_format("unknown_source")
-    <DateFormatEnum.NONE: ''>
-
-    """
-    source_lower = source.casefold()
-    if source_lower == "sources_csv":
-        return DateFormatEnum.SOURCES_CSV
-    elif source_lower == "wayback":
-        return DateFormatEnum.WAYBACK
-    else:
-        return DateFormatEnum.NONE
 
 
 # Generate a list of all available sources currently available in the package's datasources folder.
@@ -387,9 +330,9 @@ class Source:
             end_index = archive_url[0].index("/", start_index)
             # Extract the timestamp substring
             timestamp = archive_url[0][start_index:end_index]
-            output_timestamp = Source.change_datetime_format(
+            output_timestamp = td.Utils.change_datetime_format(
                 timestamp,
-                "sources_csv",
+                td.DateFormatEnum.SOURCES_CSV,
             )
             return archive_url[0], archive_url[1], output_timestamp
         except ValueError:
@@ -583,46 +526,6 @@ class Source:
             return None
 
     @staticmethod
-    def change_datetime_format(
-        input_datetime_string: str,
-        date_format_id: str,
-    ) -> str | Any:
-        """
-        Change the format of a given datetime string to a specified output format. This method takes a
-        datetime string and automatically detects its format, then converts it to the specified output format.
-        If the input string cannot be parsed, it logs an error and returns None.
-
-        Parameters
-        ----------
-        input_datetime_string : str
-            datetime string that needs to be reformatted
-
-        date_format_id : str
-            desired format for the output datetime string, following the strftime format codes.
-
-        Returns
-        -------
-           str | None
-               reformatted datetime string if successful, otherwise None
-
-        Raises
-        ------
-        ValueError
-            If the input datetime string cannot be parsed.
-
-        """
-        try:
-            # Automatically detect the format of the input datetime string
-            dt = parser.parse(input_datetime_string)
-            logger.debug(f"The datetime string has been parsed successfully: {dt}")
-            output_datetime_format = get_date_format(date_format_id)
-            output_datetime_string = dt.strftime(output_datetime_format)
-            logger.debug(f"The format is now changed to {output_datetime_format}")
-            return output_datetime_string
-        except ValueError as e:
-            raise ValueError(f"Error during datetime formatting: {e}")
-
-    @staticmethod
     def is_wayback_snapshot_available(
         input_url: str, input_timestamp: str
     ) -> str | None | Any:
@@ -653,9 +556,9 @@ class Source:
             If there is an issue with the HTTP request to the Wayback Machine API
 
         """
-        api_timestamp = Source.change_datetime_format(
+        api_timestamp = td.Utils.change_datetime_format(
             input_timestamp,
-            "wayback",
+            td.DateFormatEnum.WAYBACK,
         )
         api_url = f"http://archive.org/wayback/available?url={input_url}&timestamp={api_timestamp}"
         try:
@@ -677,9 +580,9 @@ class Source:
                 logger.info(f"Timestamp: {timestamp}")
                 logger.info(f"Status: {status}")
 
-                output_timestamp = Source.change_datetime_format(
+                output_timestamp = td.Utils.change_datetime_format(
                     timestamp,
-                    "sources_csv",
+                    td.DateFormatEnum.SOURCES_CSV,
                 )
 
                 return archived_url, output_timestamp, status
